Class RedtekBlocks was moved to the new blocks package
